Removed the default gas limit for CCM messages. Changed GasUnit type to u128 to avoid downcasting

+++ state-chain/primitives/src/lib.rs
@@ -44,7 +44,7 @@ pub type BasisPoints = u16;
 
 pub type BroadcastId = u32;
 
-pub type GasUnit = u64;
+pub type GasUnit = u128;
 
 /// The type of the Id given to threshold signature requests. Note a single request may
 /// result in multiple ceremonies, but only one ceremony should succeed.
